Performance Improvments -> Faster
The TiShadow build of the Titanium SDK does not cache CommonJS modules
loaded from the applicationDataDirectory. This is so that if an update
to the app is deployed, i.e. a file in the applicationDataDirectory is
modified, the changes will be loaded. That said loading a CommonJS
module every time that it is required makes the deployed bundle run 
slowly. So we will manage the caching of those modules in the app.